+++ packages/at_client/test/switch_atsign_test.dart
@@ -1,9 +1,29 @@
 import 'package:at_client/at_client.dart';
 import 'package:at_client/src/service/sync_service_impl.dart';
+import 'package:at_lookup/at_lookup.dart';
+import 'package:mocktail/mocktail.dart';
 import 'package:test/test.dart';
+
+class MockAtLookUp extends Mock implements AtLookUp {}
 
 void main() {
   group('A group of switch atsign tests', () {
+    test('Test that we can inject an AtLookUp instance', () async {
+      final aliceAtSign = '@alice';
+      final atClientManager = AtClientManager(aliceAtSign);
+      final alicePreference = AtClientPreference();
+      final mockAtLookUp = MockAtLookUp();
+      await atClientManager.setCurrentAtSign(
+          aliceAtSign, 'wavi', alicePreference,
+          atLookUp: mockAtLookUp);
+      expect(atClientManager.atClient.getCurrentAtSign(), aliceAtSign);
+      expect(atClientManager.atClient.getRemoteSecondary()!.atLookUp,
+          mockAtLookUp);
+      expect(
+          atClientManager.atClient.getRemoteSecondary()!.atLookUp.runtimeType,
+          MockAtLookUp);
+    });
+
     test('test switch atsign - check atsign name', () async {
       final aliceAtSign = '@alice';
       final atClientManager = AtClientManager(aliceAtSign);